Move the bytes API into its own subpackage
