+++ difflib/bytes/bytes.go
@@ -13,7 +13,7 @@
 // Getting unified diffs was the main goal of the port. Keep in mind this code
 // is mostly suitable to output text differences in a human friendly way, there
 // are no guarantees generated diffs are consumable by patch(1).
-package difflib
+package bytes
 
 import (
 	"bufio"
