+++ difflib/bytes/bytes_test.go
@@ -1,4 +1,4 @@
-package difflib
+package bytes
 
 import (
 	"bytes"
